requirements.txt
email-validator==2.1.1
flask==3.0.3
flask-login==0.6.3
flask-session==0.6.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
google-api-core==2.17.1
google-generativeai==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.7.1
sqlalchemy==2.0.28
trafilatura==1.7.0
werkzeug==3.0.1
youtube-transcript-api==0.6.2
